+++ hyperspy/tests/model/test_model.py
@@ -737,13 +737,6 @@ class TestAsSignal:
             s.data, np.array([np.zeros((2, 5)), np.ones((2, 5)) * 2])
         )
 
-    @pytest.mark.parametrize("kw", [{"parallel": True}, {"max_workers": 1}])
-    def test_warnings(self, kw):
-        with pytest.warns(
-            VisibleDeprecationWarning, match=r".* has been deprecated",
-        ):
-            _ = self.m.as_signal(**kw)
-
     def test_out_of_range_to_nan(self):
         index = 2
         self.m.channel_switches[:index] = False